requirements.txt
twitchio
pygame
requests
websockets
python-dotenv